+++ src/lua/ltable.c
@@ -320,7 +320,7 @@ static void resize (lua_State *L, Table *t, int nasize, int nhsize) {
       setobjt2t(L, luaH_set(L, t, key2tval(old)), gval(old));
   }
   if (nold != dummynode)
-    luaM_freearray(L, nold, twoto(oldhsize), Node);  /* free old array */
+    luaM_freearray(L, nold, (size_t)twoto(oldhsize), Node);  /* free old array */
 }
 
 
@@ -373,7 +373,7 @@ Table *luaH_new (lua_State *L, int narray, int nhash) {
 
 void luaH_free (lua_State *L, Table *t) {
   if (t->node != dummynode)
-    luaM_freearray(L, t->node, sizenode(t), Node);
+    luaM_freearray(L, t->node, (size_t)sizenode(t), Node);
   luaM_freearray(L, t->array, t->sizearray, TValue);
   luaM_free(L, t);
 }
